[FIX]: 오류 수정 및 테스트 검증
