+++ src/main/java/moaboa/auth/MoaboaAuthServiceApplication.java
@@ -3,7 +3,9 @@ package moaboa.auth;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @EnableFeignClients
 @SpringBootApplication
 public class MoaboaAuthServiceApplication {
