+++ src/main/java/moaboa/auth/global/BaseEntity.java
@@ -2,6 +2,7 @@ package moaboa.auth.global;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
@@ -13,6 +14,7 @@ import java.sql.Timestamp;
 import static lombok.AccessLevel.PROTECTED;
 
 @Getter
+@MappedSuperclass
 @NoArgsConstructor(access = PROTECTED)
 @EntityListeners(AuditingEntityListener.class)
 public class BaseEntity {
@@ -20,6 +22,7 @@ public class BaseEntity {
     @CreatedDate
     @Column(name = "created_at")
     private Timestamp createdAt;
+
     @LastModifiedDate
     @Column(name = "updated_at")
     private Timestamp updatedAt;
